practice project

+++ src/TaxService.java
@@ -1,5 +1,7 @@
+import java.math.BigDecimal;
+
 public class TaxService {
-    public void payOut(double taxAmount) {
+    public void payOut(BigDecimal taxAmount) {
         System.out.format("Уплачен налог в размере %.2f%n", taxAmount);
     }
 }
